Use jlong to store pointer for 64-bit Linux support

+++ trunk/pccomms/lejos/pc/comm/NXTInfo.java
@@ -4,7 +4,7 @@ public class NXTInfo {
 	public String name;
 	public String btDeviceAddress;
 	public String btResourceString;
-	public int nxtPtr;
+	public long nxtPtr;
 	public int protocol = 0;
 
 	public NXTInfo() {}
